Trying to register package dependencies within a package
This won't apply to all dependencies but some will only be useful to a specific package. For example in this case the JWT Auth package is only useful for generating auth tokens, which is only done by this package.

Other packages, such as Dispatcher are more generic, and more likely to be used by other packages we make and within apps themselves. These sorts of packages should therefore be kept in the traditional app/config/app.php file (unless it is possible to specify them more than once with no ill-effects).